refactor: package 변경 및 네이밍 변경
- dto 내부 패키지 구체화
- BlackJackGameTest 잘못된 테스트 코드 패키지 위치 수정
- CardAndScoreResult 클래스 명 변경
- BettingTable 내부 파라미터 명 변경
