+++ src/main/java/blackjack/dto/view/CardAndScoreResult.java
@@ -1,17 +1,17 @@
-package blackjack.dto;
+package blackjack.dto.view;
 
 import blackjack.domain.result.Score;
 import blackjack.domain.card.Card;
 
 import java.util.List;
 
-public class NameCardAndScore {
+public class CardAndScoreResult {
 
     private final String name;
     private final List<Card> cards;
     private final int scoreValue;
 
-    public NameCardAndScore(final String name, final List<Card> cards, final Score score) {
+    public CardAndScoreResult(final String name, final List<Card> cards, final Score score) {
         this.name = name;
         this.cards = List.copyOf(cards);
         this.scoreValue = score.getValue();
